+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
+
 ## 0.12.0 WIP
 
 ### Features
 
 - Forbids to have modules with too many imported names
-configured by `--max-imported-names` option which is 50 by default
+  configured by `--max-imported-names` option which is 50 by default
+- Forbids to raise `StopIteration` inside generators
 
 
 ## 0.11.1